requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Async & Concurrency
aiohttp==3.9.1
tenacity==8.2.3

# AI & ML
anthropic==0.18.1
openai==1.12.0  # For embeddings (API-based)
sentence-transformers==2.3.1  # Local embeddings (CPU-compatible)
huggingface-hub>=0.34.0,<1.0  # HuggingFace Inference API (compatible with transformers)
langchain==0.1.6
langchain-community==0.0.19
numpy==1.26.3
scipy==1.11.4  # Scientific computing
scikit-learn==1.4.0
pandas==2.2.0
torch==2.2.0  # Required for sentence-transformers (CPU version)
transformers>=4.36.0  # Fix compatibility with torch 2.2+

# Observability & Telemetry
opentelemetry-api==1.23.0
opentelemetry-sdk==1.23.0
opentelemetry-instrumentation-fastapi==0.44b0
opentelemetry-exporter-otlp==1.23.0
opentelemetry-instrumentation-asyncio==0.44b0
opentelemetry-exporter-prometheus==0.44b0
prometheus-client==0.19.0

# Database & Storage
psycopg[binary,pool]==3.1.17  # PostgreSQL/TimescaleDB
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.25
alembic==1.13.1

# Vector Database
qdrant-client==1.7.3
chromadb==0.4.22

# Caching & Queue
redis[hiredis]==5.0.1
celery==5.3.6

# Utilities
python-dotenv==1.0.1
structlog==24.1.0
python-json-logger==2.0.7
httpx==0.26.0

# Circuit Breaker & Resilience
pybreaker==1.0.1

